🧹 chore: 使用uv管理项目；使用cz提交；使用uv_build构建包
