+++ pyproject.toml
@@ -1,0 +1,129 @@
+[project]
+name = "nonebot-plugin-impart"
+version = "0.1.2.5.3"
+description = "NoneBot2 Impart Plugin Plus"
+readme = "README.md"
+authors = [{ name = "YuuzukiRin", email = "yuuzukirin@outlook.com" }]
+urls = { Repository = "https://github.com/YuuzukiRin/nonebot_plugin_impart" }
+requires-python = ">=3.10,<3.14"
+dependencies = [
+  "nonebot2>=2.4.3,<3.0.0",
+  "nonebot-plugin-localstore>=0.7.4,<1.0.0",  # 存储文件
+  "nonebot-plugin-apscheduler>=0.5.0,<1.0.0", # 定时任务
+  "nonebot-plugin-uninfo>=0.9.0,<1.0.0",   # 多平台用户信息获取
+  "nonebot-plugin-alconna>=0.59.3,<1.0.0", # 跨平台命令解析拓展
+  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
+  "Pillow>=10.2.0",
+  "httpx>=0.27.0",
+  "sqlalchemy>=2.0.20",
+  "aiosqlite>=0.19.0",
+]
+license = { text = "MIT" }
+
+
+[dependency-groups]
+dev = [
+  "commitizen>=4.8.3",
+  "cz-conventional-gitmoji>=0.7.0",
+  "nonebot2[fastapi]>=2.4.2,<3.0.0",
+  "pre-commit>=4.3.0",
+  "ruff>=0.12.1,<1.0.0",
+]
+
+test = [
+  "nonebot2[fastapi]>=2.4.2,<3.0.0",
+  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
+  "nonebug>=0.3.7,<1.0.0",
+  "pytest-xdist>=3.8.0,<4.0.0",
+  "pytest-asyncio>=1.1.0,<1.2.0",
+]
+
+[tool.nonebot]
+plugins = ["nonebot_plugin_template"]
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "session"
+pythonpath = ["src"]
+addopts = [
+  "-v",               # 详细输出
+  "-s",               # 显示打印信息
+  "--tb=short",       # 简短的错误回溯
+  "-ra",              # 显示所有测试结果摘要
+  "--strict-markers", # 严格标记模式
+  "--import-mode=prepend", # 导入模式
+]
+
+[tool.ruff]
+line-length = 120
+target-version = "py313"
+
+[tool.ruff.format]
+line-ending = "lf"
+quote-style = "double"
+indent-style = "space"
+
+[tool.ruff.lint]
+select = [
+  "F",     # Pyflakes
+  "W",     # pycodestyle warnings
+  "E",     # pycodestyle errors
+  "I",     # isort
+  "UP",    # pyupgrade
+  "ASYNC", # flake8-async
+  "C4",    # flake8-comprehensions
+  "T10",   # flake8-debugger
+  "T20",   # flake8-print
+  "PYI",   # flake8-pyi
+  "PT",    # flake8-pytest-style
+  "TID",   # flake8-tidy-imports
+  "RUF",   # Ruff-specific rules
+]
+
+ignore = [
+  "UP037",  # quoted-annotation
+  "RUF001", # ambiguous-unicode-character-string
+  "RUF002", # ambiguous-unicode-character-docstring
+  "RUF003", # ambiguous-unicode-character-comment
+]
+
+
+[tool.ruff.lint.isort]
+force-sort-within-sections = true
+known-first-party = ["tests/*"]
+extra-standard-library = ["typing_extensions"]
+
+[tool.ruff.lint.flake8-pytest-style]
+fixture-parentheses = false
+mark-parentheses = false
+
+[tool.ruff.lint.pyupgrade]
+keep-runtime-typing = true
+
+
+[tool.pyright]
+pythonVersion = "3.13"
+pythonPlatform = "All"
+defineConstant = { PYDANTIC_V2 = true }
+executionEnvironments = [
+  { root = "./tests", extraPaths = [
+    "./src",
+  ] },
+  { root = "./src" },
+]
+typeCheckingMode = "standard"
+reportShadowedImports = false
+disableBytesTypePromotions = true
+
+
+[build-system]
+requires = ["uv_build>=0.8.0,<0.9.0"]
+build-backend = "uv_build"
+
+
+[tool.commitizen]
+name = "cz_gitmoji"
+tag_format = "v$version"
+version_scheme = "pep440"
+version_provider = "uv"
+update_changelog_on_bump = true
